% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer_tab.R
\name{cramer_tab}
\alias{cramer_tab}
\title{get_cramer}
\usage{
cramer_tab(var = NULL, metric = "cramer", data = NULL, weight = NULL)
}
\arguments{
\item{var}{A list of variable name you want to compute "metric" on}

\item{metric}{Choose between 'Cramer' or 'YuleQ' parameter to select one of this metrics}

\item{data}{The data table you want to compute 'metric' parameter on}

\item{weight}{The weight vector if there is one in the dataframe}
}
\value{
A list of gt information
}
\description{
Get Cramer or YuleQ's tab
}
\examples{
data <- graphPAF::Hordaland_data
cramer_tab(c("y", "urban.rural"), metric = 'cramer', data = data)
}
