% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AAF.R
\name{get_AAF}
\alias{get_AAF}
\title{get_AAF}
\usage{
get_AAF(
  model = NULL,
  nvar = NULL,
  name_var = NULL,
  vars_dep = NULL,
  data = NULL,
  nbootstrap = 50,
  nperm = 100
)
}
\arguments{
\item{model}{The model you want to compute AAF on}

\item{nvar}{You can choose this \code{integer} parameter if you prefer to add explanatory variables in the order they appear in the model. You don't have to use \code{name_var} parameter.}

\item{name_var}{A list of explanatory variables you want to compute AAF on}

\item{vars_dep}{A string with the name of the variable you want to explain}

\item{data}{The data you used to fit the model (with the variable in name_var and vars_dep)}

\item{nbootstrap}{An integer with the n bootsrap replication you want. By default, 50.}

\item{nperm}{Number of permutation between modalities and variable to compute the AAF}
}
\value{
A sf table with the result of the AAF computation
}
\description{
Get the AAF
}
\examples{
set.seed(9876)
data <- data.frame(
   "var_dep" = as.factor(sample(c(rep_len(c(0, 1), length.out = 1000)), size = 100)),
   "var_1" = as.factor(sample(c(rep_len(c(1, 2, 3), length.out = 1000)), size = 100)),
   "var_2" = as.factor(sample(c(rep_len(c(3, 1, 2), length.out = 100)), size = 100))
   )
glmmodel <- glm(var_dep ~ var_1 + var_2,
                data = data, family = binomial("logit"))
\donttest{
res <- get_AAF(model = glmmodel, nvar = 3, vars_dep = "var_dep", data = data,
               nbootstrap = 5)
res <- get_AAF(model = glmmodel,
               name_var = c("var_1", "var_2"),
              vars_dep = "var_dep", data = data, nbootstrap = 5)
}
}
