% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer_tab.R
\name{get_list}
\alias{get_list}
\title{Title}
\usage{
get_list(name_var)
}
\arguments{
\item{name_var}{A list of variable you want to pass a name for a list}
}
\value{
a list of zero named after the variable in name_var parameter
}
\description{
Title
}
\examples{
vec <- get_list(name_var = c("var1", "var2"))
}
