% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer_tab.R
\name{cramer_tab}
\alias{cramer_tab}
\title{Title}
\usage{
cramer_tab(var, metric = "cramer", data)
}
\arguments{
\item{var}{A list of variable name you want to compute "metric" on}

\item{metric}{Choose between 'Cramer' or 'YuleQ' parameter to select one of this metrics}

\item{data}{The data table you want to compute 'metric' parameter on}
}
\value{
A list of gt information
}
\description{
Title
}
\examples{
cramer_tab(c("var_1", "var_2"), metric = 'cramer', data = data)
}
