% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AAF.R
\name{get_AAF}
\alias{get_AAF}
\title{Get the AAF}
\usage{
get_AAF(
  model = NULL,
  nvar = NULL,
  name_var = NULL,
  vars_dep = NULL,
  data = NULL,
  nbootstrap = 50
)
}
\arguments{
\item{model}{The model you want to compute AAF on}

\item{nvar}{You can choose this \code{integer} parameter if you prefer to add explanatory variables in the order they appear in the model. You don't have to use \code{name_var} parameter.}

\item{name_var}{A list of explanatory variables you want to compute AAF on}

\item{vars_dep}{A string with the name of the variable you want to explain}

\item{data}{The data you used to fit the model (with the variable in name_var and vars_dep)}

\item{nbootstrap}{An integer with the n bootsrap replication you want. By default, 50.}
}
\value{
A sf table with the result of the AAF computation
}
\description{
Get the AAF
}
\examples{
\dontrun{res <- get_AAF(model = glmmodel, nvar = 5, vars_dep = "dependant_var", data = data)}
\dontrun{res <- get_AAF(model = glmmodel, name_var = c("var_1", "var_2", "var_3", "var_4", "var_5"),
                        vars_dep = "dependant_var", data = data, nbootstrap = 100)}
}
