% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yuleq.R
\name{get_yuleq}
\alias{get_yuleq}
\title{Get Yule's Q}
\usage{
get_yuleq(
  vars_dep = NULL,
  name_var = c(),
  data = NULL,
  source = "",
  alpha = 0.01
)
}
\arguments{
\item{vars_dep}{A string with name of the dependant variable to use as one variable for the 2x2 table}

\item{name_var}{A vector of explanatory variable you want to compute Yule's Q on}

\item{data}{A dataframe with the variable named before}

\item{source}{A string with the name of the data source}

\item{alpha}{A integer with the alpha value for fisher.test and IC}
}
\value{
A gt table with the Yule's computed for the variable in name_var
}
\description{
Get Yule's Q
}
\examples{
\dontrun{get_yuleq(vars_dep = "dependant_var", name_var = c("var_1", "var_2"),
                    data = data, source = "dataSource - Year", alpha = 0.01)}
}
