% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_tab.R
\name{result_tab}
\alias{result_tab}
\title{Title}
\usage{
result_tab(model, var_ref, res_AAF, data)
}
\arguments{
\item{model}{The model used to perform statistical analysis}

\item{var_ref}{The variable to explain in the statistical model}

\item{res_AAF}{the result of the AAF computation with \code{get_AAF()} function (done outside of this function because of the length of the computation)}

\item{data}{The data used for the statistical analysis}
}
\value{
A gt table
}
\description{
Title
}
\examples{
res_tab <- result_tab(model, var_ref = "var_0", res_AAF = resAAF$res, data = data)
}
